+++ src/commands/mod.rs
@@ -1,6 +1,6 @@
-mod math;
-mod info;
-mod neko;
-mod setup;
-mod settings;
-mod utilities;
+pub mod math;
+pub mod info;
+pub mod neko;
+pub mod setup;
+pub mod settings;
+pub mod utilities;
