+++ src/main.rs
@@ -3,24 +3,24 @@ use reqwest;
 use sqlx::SqlitePool;
 use std::env;
 
+mod commands;
+
+
+use crate::commands::{
+    info::*,
+    math::*,
+    neko::*,
+    setup::*,
+    settings::*,
+    utilities::*
+};
+
 struct Data {
     reqwest: reqwest::Client,
     sqlite: SqlitePool,
 } // User data, which is stored and accessible in all command invocations
 type Error = Box<dyn std::error::Error + Send + Sync>;
 type Context<'a> = poise::Context<'a, Data, Error>;
-
-/// Displays your or another user's account creation date
-#[poise::command(slash_command, prefix_command)]
-async fn age(
-    ctx: Context<'_>,
-    #[description = "Selected user"] user: Option<serenity::User>,
-) -> Result<(), Error> {
-    let u = user.as_ref().unwrap_or_else(|| ctx.author());
-    let response = format!("{}'s account was created at {}", u.name, u.created_at());
-    ctx.say(response).await?;
-    Ok(())
-}
 
 #[tokio::main]
 async fn main() {
@@ -46,7 +46,9 @@ async fn main() {
 
     let framework = poise::Framework::builder()
         .options(poise::FrameworkOptions {
-            commands: vec![age()],
+            commands: vec![
+                
+            ],
             ..Default::default()
         })
         .setup(|ctx, _ready, framework| {
@@ -57,7 +59,8 @@ async fn main() {
                     sqlite: database,
                 })
             })
-        }).build();
+        })
+        .build();
 
     let client = serenity::ClientBuilder::new(token, intents)
         .framework(framework)
